Refactor person package to card package
