+++ src/main/java/seedu/address/logic/Logic.java
@@ -8,7 +8,7 @@ import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.ReadOnlyDeck;
-import seedu.address.model.person.Card;
+import seedu.address.model.card.Card;
 
 /**
  * API of the Logic component
