+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -8,7 +8,7 @@ import seedu.address.commons.core.Messages;
 import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
-import seedu.address.model.person.Card;
+import seedu.address.model.card.Card;
 
 /**
  * Deletes a card identified using it's displayed index from the address book.
