+++ src/main/java/seedu/address/model/Deck.java
@@ -5,8 +5,8 @@ import static java.util.Objects.requireNonNull;
 import java.util.List;
 
 import javafx.collections.ObservableList;
-import seedu.address.model.person.Card;
-import seedu.address.model.person.UniquePersonList;
+import seedu.address.model.card.Card;
+import seedu.address.model.card.UniquePersonList;
 
 /**
  * Wraps all data at the address-book level
