+++ src/main/java/seedu/address/model/Model.java
@@ -6,7 +6,7 @@ import java.util.function.Predicate;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.index.Index;
-import seedu.address.model.person.Card;
+import seedu.address.model.card.Card;
 
 /**
  * The API of the Model component.
