+++ src/main/java/seedu/address/model/ModelManager.java
@@ -12,7 +12,7 @@ import javafx.collections.transformation.FilteredList;
 import seedu.address.commons.core.GuiSettings;
 import seedu.address.commons.core.LogsCenter;
 import seedu.address.commons.core.index.Index;
-import seedu.address.model.person.Card;
+import seedu.address.model.card.Card;
 import seedu.address.model.powerdeck.PowerDeck;
 
 /**
