+++ src/main/java/seedu/address/model/powercard/UniqueCardList.java
@@ -8,8 +8,8 @@ import java.util.List;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
-import seedu.address.model.person.exceptions.DuplicatePersonException;
-import seedu.address.model.person.exceptions.PersonNotFoundException;
+import seedu.address.model.card.exceptions.DuplicatePersonException;
+import seedu.address.model.card.exceptions.PersonNotFoundException;
 
 /**
  * A list of cards that enforces uniqueness between its elements and does not allow nulls.
