+++ src/main/java/seedu/address/model/util/SampleDataUtil.java
@@ -6,11 +6,11 @@ import java.util.stream.Collectors;
 
 import seedu.address.model.Deck;
 import seedu.address.model.ReadOnlyDeck;
-import seedu.address.model.person.Address;
-import seedu.address.model.person.Card;
-import seedu.address.model.person.Email;
-import seedu.address.model.person.Name;
-import seedu.address.model.person.Phone;
+import seedu.address.model.card.Address;
+import seedu.address.model.card.Card;
+import seedu.address.model.card.Email;
+import seedu.address.model.card.Name;
+import seedu.address.model.card.Phone;
 import seedu.address.model.tag.Tag;
 
 /**
