+++ src/main/java/seedu/address/storage/JsonSerializableAddressBook.java
@@ -11,7 +11,7 @@ import com.fasterxml.jackson.annotation.JsonRootName;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.model.Deck;
 import seedu.address.model.ReadOnlyDeck;
-import seedu.address.model.person.Card;
+import seedu.address.model.card.Card;
 
 /**
  * An Immutable Deck that is serializable to JSON format.
