+++ src/main/java/seedu/address/ui/PersonCard.java
@@ -7,7 +7,7 @@ import javafx.scene.control.Label;
 import javafx.scene.layout.FlowPane;
 import javafx.scene.layout.HBox;
 import javafx.scene.layout.Region;
-import seedu.address.model.person.Card;
+import seedu.address.model.card.Card;
 
 /**
  * An UI component that displays information of a {@code Card}.
