+++ src/main/java/seedu/address/ui/PersonListPanel.java
@@ -8,7 +8,7 @@ import javafx.scene.control.ListCell;
 import javafx.scene.control.ListView;
 import javafx.scene.layout.Region;
 import seedu.address.commons.core.LogsCenter;
-import seedu.address.model.person.Card;
+import seedu.address.model.card.Card;
 
 /**
  * Panel containing the list of persons.
