+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -26,7 +26,7 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.ReadOnlyDeck;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.person.Card;
+import seedu.address.model.card.Card;
 import seedu.address.storage.JsonAddressBookStorage;
 import seedu.address.storage.JsonUserPrefsStorage;
 import seedu.address.storage.StorageManager;
