+++ src/test/java/seedu/address/logic/commands/AddCommandTest.java
@@ -21,7 +21,7 @@ import seedu.address.model.Deck;
 import seedu.address.model.Model;
 import seedu.address.model.ReadOnlyDeck;
 import seedu.address.model.ReadOnlyUserPrefs;
-import seedu.address.model.person.Card;
+import seedu.address.model.card.Card;
 import seedu.address.testutil.PersonBuilder;
 
 public class AddCommandTest {
