+++ src/test/java/seedu/address/logic/commands/CommandTestUtil.java
@@ -17,8 +17,8 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Deck;
 import seedu.address.model.Model;
-import seedu.address.model.person.Card;
-import seedu.address.model.person.NameContainsKeywordsPredicate;
+import seedu.address.model.card.Card;
+import seedu.address.model.card.NameContainsKeywordsPredicate;
 import seedu.address.testutil.EditPersonDescriptorBuilder;
 
 /**
