+++ src/test/java/seedu/address/logic/commands/EditCommandTest.java
@@ -23,7 +23,7 @@ import seedu.address.model.Deck;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
-import seedu.address.model.person.Card;
+import seedu.address.model.card.Card;
 import seedu.address.testutil.EditPersonDescriptorBuilder;
 import seedu.address.testutil.PersonBuilder;
