+++ src/test/java/seedu/address/testutil/AddressBookBuilder.java
@@ -1,7 +1,7 @@
 package seedu.address.testutil;
 
 import seedu.address.model.Deck;
-import seedu.address.model.person.Card;
+import seedu.address.model.card.Card;
 
 /**
  * A utility class to help with building Addressbook objects.
