+++ src/test/java/seedu/address/testutil/TypicalPersons.java
@@ -16,7 +16,7 @@ import java.util.Arrays;
 import java.util.List;
 
 import seedu.address.model.Deck;
-import seedu.address.model.person.Card;
+import seedu.address.model.card.Card;
 
 /**
  * A utility class containing a list of {@code Card} objects to be used in tests.
